requirements.txt
fastapi==0.78.0
uvicorn==0.18.2
jinja2==3.1.2
pillow==9.2.0
starlette==0.23.1
